+++ src/Executor/Commands/SOB.cs
@@ -1,4 +1,5 @@
 using Executor.Arguments;
+using Executor.Arguments.Abstraction;
 using Executor.CommandTypes;
 using Executor.Memories;
 using Executor.States;
@@ -30,5 +31,5 @@ public class SOB : BaseCommand
     {
     }
 
-    public override ushort Opcode => (ushort)Convert.ToUInt16("077000", 8);
+    public override ushort Opcode => Convert.ToUInt16("077000", 8);
 }
